use cargo-eval to build wix installer
